Remove SAGESharp.SLB.Conversation namespace
This is part of a cleanup to add classes for all SLB
files, having this complicated structure for namespaces
is not necessary
