+++ SAGESharpTests/SAGESharpTests.csproj
@@ -95,18 +95,18 @@
     <Compile Include="SLB\IO\FooterEntryTests.cs" />
     <Compile Include="SLB\IO\SLBFooterWriterTests.cs" />
     <Compile Include="SLB\IO\StringBinaryWriterTests.cs" />
-    <Compile Include="SLB\Level\Conversation\CharacterEqualityTests.cs" />
-    <Compile Include="SLB\Level\Conversation\Defaults.cs" />
-    <Compile Include="SLB\Level\Conversation\FrameEqualityTests.cs" />
-    <Compile Include="SLB\Level\Conversation\InfoEqualityTests.cs" />
-    <Compile Include="SLB\Level\Conversation\IO\CharacterBinaryWriterTests.cs" />
-    <Compile Include="SLB\Level\Conversation\IO\ConversationBinaryAccessorTests.cs" />
-    <Compile Include="SLB\Level\Conversation\IO\ConversationBinaryWriterTests.cs" />
-    <Compile Include="SLB\Level\Conversation\IO\ConversationFooterGeneratorTests.cs" />
-    <Compile Include="SLB\Level\Conversation\IO\FrameBinaryWriterTests.cs" />
-    <Compile Include="SLB\Level\Conversation\IO\InfoBinaryWriterTests.cs" />
-    <Compile Include="SLB\Level\Conversation\SerializationTests.cs" />
-    <Compile Include="SLB\Level\Conversation\TestData.cs" />
+    <Compile Include="SLB\Level\CharacterEqualityTests.cs" />
+    <Compile Include="SLB\Level\Defaults.cs" />
+    <Compile Include="SLB\Level\FrameEqualityTests.cs" />
+    <Compile Include="SLB\Level\InfoEqualityTests.cs" />
+    <Compile Include="SLB\Level\IO\CharacterBinaryWriterTests.cs" />
+    <Compile Include="SLB\Level\IO\ConversationBinaryAccessorTests.cs" />
+    <Compile Include="SLB\Level\IO\ConversationBinaryWriterTests.cs" />
+    <Compile Include="SLB\Level\IO\ConversationFooterGeneratorTests.cs" />
+    <Compile Include="SLB\Level\IO\FrameBinaryWriterTests.cs" />
+    <Compile Include="SLB\Level\IO\InfoBinaryWriterTests.cs" />
+    <Compile Include="SLB\Level\SerializationTests.cs" />
+    <Compile Include="SLB\Level\TestData.cs" />
     <Compile Include="SLB\Point3DEqualityTests.cs" />
     <Compile Include="Testing\AbstractTestCaseData.cs" />
     <Compile Include="Testing\Matchers.cs" />
