+++ SAGESharpTests/SLB/Level/Defaults.cs
@@ -3,11 +3,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-using SAGESharp.SLB;
-using SAGESharp.SLB.Level.Conversation;
 using System.Collections.Generic;
 
-namespace SAGESharp.SLB.Level.Conversation
+namespace SAGESharp.SLB.Level
 {
     static class Defaults
     {
