+++ SAGESharpTests/SLB/Level/IO/ConversationBinaryWriterTests.cs
@@ -11,7 +11,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 
-namespace SAGESharp.SLB.Level.Conversation.IO
+namespace SAGESharp.SLB.Level.IO
 {
     class ConversationBinaryWriterTests
     {
