+++ SAGESharpTests/SLB/Level/IO/ConversationFooterGeneratorTests.cs
@@ -10,7 +10,7 @@ using SAGESharp.Testing;
 using System;
 using System.Collections.Generic;
 
-namespace SAGESharp.SLB.Level.Conversation.IO
+namespace SAGESharp.SLB.Level.IO
 {
     class ConversationFooterGeneratorTests
     {
