+++ SAGESharpTests/SLB/Level/IO/InfoBinaryWriterTests.cs
@@ -12,7 +12,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 
-namespace SAGESharp.SLB.Level.Conversation.IO
+namespace SAGESharp.SLB.Level.IO
 {
     class InfoBinaryWriterTests
     {
